+++ grid2op/Observation/baseObservation.py
@@ -4744,3 +4744,54 @@ class BaseObservation(GridObjects):
         
         # update the was_alert_used_after_attack !
         self.was_alert_used_after_attack[:] = env._was_alert_used_after_attack
+        
+    def get_back_to_ref_state(
+        self,
+        storage_setpoint: float=0.5,
+        precision: int=5,
+    ) -> Dict[Literal["powerline",
+                      "substation",
+                      "redispatching",
+                      "storage",
+                      "curtailment"],
+              List["grid2op.Action.BaseAction"]]:
+        """
+        Allows to retrieve the list of actions that needs to be performed
+        to get back the grid in the "reference" state (all elements connected
+        to busbar 1, no redispatching, no curtailment)
+        
+        
+        .. versionadded:: 1.9.9
+        
+        This function uses the method of the underlying action_space used 
+        for the forecasts.
+        
+        See :func:`grid2op.Action.SerializableActionSpace.get_back_to_ref_state`
+        for more information.
+        
+        Examples
+        --------
+        
+        You can use it like this:
+        
+        .. code-block:: python
+        
+            import grid2op
+
+            env_name = "l2rpn_case14_sandbox"
+            env = grid2op.make(env_name)
+            obs = env.reset(seed=1)
+
+            # perform a random action
+            obs, reward, done, info = env.step(env.action_space.sample())
+            assert not done # you might end up in a "done" state depending on the random action
+            
+            acts = obs.get_back_to_ref_state()
+            print(acts)
+        """     
+        if self.action_helper is None:
+            raise Grid2OpException("Trying to use this function when no action space is "
+                                   "is available.")   
+        if self._is_done:
+            raise Grid2OpException("Cannot use this function in a 'done' state.")
+        return self.action_helper.get_back_to_ref_state(self, storage_setpoint, precision)